+++ src/interfaces.ts
@@ -1,6 +1,4 @@
 export interface MarkdownTableFormatterSettings {
-    formatOnSave: boolean;
-    autoSelectEntireDocument: boolean;
     spacePadding: number;
     keepFirstAndLastPipes: boolean;
     defaultTableJustification: string;