+++ src/table-formatter.ts
@@ -8,8 +8,6 @@ function getSettings(): MarkdownTableFormatterSettings {
     let vscodeConfig = vscode.workspace.getConfiguration('markdown-table-formatter');
     // Forcing cast because defaults are defined in packages.json, so always have a value
     return {
-        formatOnSave: <boolean>vscodeConfig.get<boolean>('formatOnSave'),
-        autoSelectEntireDocument: <boolean>vscodeConfig.get<boolean>('autoSelectEntireDocument'),
         spacePadding: <number>vscodeConfig.get<number>('spacePadding'),
         keepFirstAndLastPipes: <boolean>vscodeConfig.get<boolean>('keepFirstAndLastPipes'),
         defaultTableJustification: <string>vscodeConfig.get<string>('defaultTableJustification'),
@@ -18,69 +16,44 @@ function getSettings(): MarkdownTableFormatterSettings {
     };
 }
 
-export class TableFormatter {
-    public format(editor: vscode.TextEditor, force: boolean = false) {
-        const emptySelection = editor.selections.every(s => s.isEmpty);
-        if (!getSettings().markdownGrammarScopes.includes(editor.document.languageId)) {
-            return undefined;
-        }
-        if (force || (emptySelection && getSettings().autoSelectEntireDocument)) {
-            let tables: any[] = this.tablesIn(editor.document);
-            editor.edit(editBuilder => {
-                tables.forEach(table => {
-                    editBuilder.replace(table.range, formatTable(table.match, getSettings()));
-                });
-            });
-        } else {
-            let tables: any[] = this.tablesIn(editor.document, editor.selections);
-            editor.edit(editBuilder => {
-                tables.forEach(table => {
-                    editBuilder.replace(table.range, formatTable(table.match, getSettings()));
-                });
-            });
-        }
+export class MarkdownTableFormatterProvider implements vscode.DocumentFormattingEditProvider, vscode.DocumentRangeFormattingEditProvider {
+
+    provideDocumentFormattingEdits(document: vscode.TextDocument, options: vscode.FormattingOptions, token: vscode.CancellationToken): vscode.ProviderResult<vscode.TextEdit[]> {
+        let fullDocumentRange = document.validateRange(new vscode.Range(0, 0, document.lineCount + 1, 0));
+        return this.formatDocument(document, fullDocumentRange);
     }
 
-    public formatDocument(documento: vscode.TextDocument): vscode.TextEdit[] {
+    provideDocumentRangeFormattingEdits(document: vscode.TextDocument, range: vscode.Range, options: vscode.FormattingOptions, token: vscode.CancellationToken): vscode.ProviderResult<vscode.TextEdit[]> {
+        return this.formatDocument(document, range);
+    }
+
+    private formatDocument(document: vscode.TextDocument, range: vscode.Range): vscode.TextEdit[] {
         let edits: vscode.TextEdit[] = [];
-        if (!getSettings().markdownGrammarScopes.includes(documento.languageId)) {
+        // This check is in case some grammar is removed and VSCode is not reloaded yet
+        if (!getSettings().markdownGrammarScopes.includes(document.languageId)) {
+            vscode.window.showWarningMessage(`Markdown table formatter is not enabled for '${document.languageId}' language!`);
             return edits;
         }
-        let tables: any[] = this.tablesIn(documento);
+        let tables: any[] = this.tablesIn(document, range);
         tables.forEach(table => {
             edits.push(new vscode.TextEdit(table.range, formatTable(table.match, getSettings())));
         });
         return edits;
     }
 
-    private tablesIn(document: vscode.TextDocument, forRanges: vscode.Range[] = []) {
+    private tablesIn(document: vscode.TextDocument, range: vscode.Range) {
         var items: any = [];
-        // Think in a way to optimize this
-        if (forRanges.length === 0) {
-            const firstLine = document.lineAt(0);
-            const lastLine = document.lineAt(document.lineCount - 1);
-            const textRange = new vscode.Range(0,
-                firstLine.range.start.character,
-                document.lineCount - 1,
-                lastLine.range.end.character);
-            forRanges.push(textRange);
+
+        const text = document.getText(range);
+        var matches;
+        while ((matches = tableRegex.exec(text)) !== null) {
+            let offset = document.offsetAt(range.start);
+            let start = document.positionAt(offset+matches.index);
+            let text = matches[0].trim();
+            let end = document.positionAt(offset+matches.index+text.length);
+            let nrange = new vscode.Range(start, end);
+            items.push({ match: matches, range: nrange });
         }
-        forRanges.forEach(range => {
-            const text = document.getText();
-            let match = tableRegex.exec(text);
-            while (match !== null) {
-                if (match) {
-                    let start = document.positionAt(match.index);
-                    let end = document.positionAt(match.index + match[0].length);
-                    let nrange = new vscode.Range(start, end);
-                    let r = nrange.intersection(range);
-                    if (r) {
-                        items.push({ match: match, range: r });
-                    }
-                }
-                match = tableRegex.exec(text);
-            }
-        });
         return items;
     }
 }